requirements.txt
streamlit
pandas
streamlit-gsheets-connection
google-auth
gspread
selenium